benchmarks/pyproject.toml
[tool.poetry]
name = "lion-benchmarks"
version = "0.0.0"
description = ""
authors = ["Michael P. Jung <michael.jung@terreon.de>"]
license = "BSD-2-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.5"
marshmallow = "^3.6.0"
serpy = "^0.3.1"
django-rest-framework = "^0.1.0"
pyparsing = "^2.4.7"
django-rest-marshmallow = "^4.0.2"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
